requirements.txt
torch>=1.13
torchvision>=0.14
numpy
Pillow
tqdm
scikit-learn
pandas
matplotlib
opencv-python
